@ V1.4.0：第5次补充需求修复-湿度映射、设置步进、报警启停与ESP8266连接优化
1. 湿度映射修正：soil_adc_to_percent恢复反转映射（ADC 0→100%湿润，ADC 4095→0%干燥），匹配电阻式传感器实际特性。
2. 设置步进改为±1%：湿度上下限每次调整1%（原5%），温度步进保持±1°C不变。
3. K4报警手动启停：改为toggle开关，按一次静音、再按恢复，移除60秒静音超时机制。
4. ESP8266 UART TX超时优化：HAL_UART_Transmit超时从1000-2000ms降至50ms，解决继电器EMI导致主循环阻塞丢按键。
5. MQTT连接响应修正：AT+MQTTCONN等待+MQTTCONNECTED事件（原等OK），避免"假连接"导致数据上报全部超时。
6. MQTT连接稳定化：新增ESP_CONN_MQTT_WAIT状态，MQTTCONN成功后延时500ms再订阅。
7. 订阅容错机制：SUB1/SUB2各最多重试2次，仍失败则跳过继续运行，不阻断数据上报。
8. 首次发布立即触发：进入MQTT_OK后g_pub_tick=0，秒级完成首次数据上报。
9. OLED连接状态细分：10种状态显示（W:Rs/W:AT/W:Jn/M:Cf/M:Cn/M:S1/M:S2/M:OK/Err/W:--），无需串口即可定位连接问题。
10. 新增README.md与文档技能配置。

@

+++ stm32-code/User/Inc/data_store.h
@@ -10,8 +10,8 @@ extern "C" {
 typedef struct
 {
   uint8_t  plant_type;
-  uint16_t soil_low;
-  uint16_t soil_high;
+  uint8_t  soil_low;
+  uint8_t  soil_high;
   uint8_t  temp_low;
   uint8_t  temp_high;
 } threshold_config_t;
